+++ cucaibafront/src/components/DetalleAgente.jsx
@@ -1,7 +1,6 @@
 import React from "react";
-import { Navigate, redirect, useParams } from "react-router-dom";
+import { Navigate, useParams } from "react-router-dom";
 import "../assets/styles/detalle.css";
-import BackButton from "../components/UI/BackButton";
 import { useAgentes } from "../hooks/useAgentes";
 import Spinner from "./UI/Spinner";
 import { GoTriangleDown } from "react-icons/go";
