+++ cucaibafront/src/components/PostOperativos.jsx
@@ -79,11 +79,11 @@ const PostOperativos = () => {
             value={operativo.referencia}
             onChange={(e) => {
               const newValue = e.target.value;
-                  if (newValue >= 0) {
-              setOperativo({ ...operativo, referencia: e.target.value });
-              validateString(e.target.name, e.target.value);
+              if (newValue >= 0) {
+                setOperativo({ ...operativo, referencia: e.target.value });
+                validateString(e.target.name, e.target.value);
+              }
             }}
-          }
           />
           {showError.referencia && (
             <div style={{ color: "red" }}>
