+++ cucaibafront/src/components/styles/CrearModulo.css
@@ -1,0 +1,32 @@
+.container {
+  padding: 20px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+  border: 1px solid gray;
+  min-width: calc(20vw - 20px);
+  border-radius: 15px;
+}
+
+.modulo {
+  display: flex;
+  justify-content: flex-start;
+  margin-left: 0;
+  color: #5dade2;
+}
+
+.hrstyle {
+  margin: 10px 0;
+  border: none;
+  border-bottom: 5px solid #5dade2;
+}
+
+.hrstyle2 {
+    margin-top: 4rem;
+    border: none;
+    border-bottom: 5px solid #5DADE2;
+  }
+
+.spanObligatorio {
+    color: red; 
+    margin-left: 5px; 
+    font-size: 20px;
+}
