+++ cucaibafront/src/components/styles/Liquidaciones.css
@@ -1,0 +1,8 @@
+.checkbox {
+    width: 20px;
+    height: 20px;
+}
+
+.spanObligatorio {
+    color: red;
+}
