+++ cucaibafront/src/components/styles/Modulos.css
@@ -1,0 +1,18 @@
+.acciones {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+    max-width: 50%;
+    margin-left: 1rem;
+}
+
+.form-modulo {
+    position: fixed;
+    top: 45%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    background-color: white;
+    z-index: 999;
+    box-shadow: 0px 0px 48px 100vw rgba(0, 0, 0, 0.29);
+    border-radius: 15px;
+}
