+++ cucaibafront/src/components/styles/PostAgentes.css
@@ -1,0 +1,4 @@
+.spanObligatorio {
+    color: "red"; 
+    display: "none" 
+}
